[sonic-fips]: Use curl instead of wget when downloading files
When using wget to download files, if the download fails, then this
still leaves a zero-length file in the destination, and on a retry, this
task is assumed to have passed. The overall build might fail later when
building the image itself. There are some workaround available, but
instead of making a code a bit messy, use curl instead, which has an
option to not create a file if the download fails.

Signed-off-by: Saikrishna Arcot <sarcot@microsoft.com>