Remove all version control files
Between slave container files that are no longer used and version
updates that are coming with the Trixie upgrade, the version control
files appear to be resulting in mismatching versions of Python wheels
between the slave container and target image.

Clear the slate completely and use the latest versions for now. Later,
the version control files will get recreated.

Signed-off-by: Saikrishna Arcot <sarcot@microsoft.com>