Use Rust installed from Debian instead of from rustup
Debian Trixie has Rust 1.85, which is slightly older than the Rust 1.86
that would've been installed in Bookworm, but should work just as well.
Use Rust 1.85 from Trixie so that we can get as much as possible from
Debian.

For Debian Bookworm, downgrade from 1.86 to 1.85 to match what will be
in Trixie.

Signed-off-by: Saikrishna Arcot <sarcot@microsoft.com>