amend! Remove all version control files
Recreate all version control files from scratch

With the trixie upgrade, all of the package versions for the base image
will have changed, meaning the version control files will not be useful
at all for the base image.

Take this opportunity to recreate all of the version files (including
the ones for the containers).

Signed-off-by: Saikrishna Arcot <sarcot@microsoft.com>